requirements.txt
pyusb==1.0.0
bottle==0.12.13
paste==2.0.3
#git+https://github.com/atlas0fd00m/rfcat.git#egg=rflib&subdirectory=rflib
#git+https://github.com/atlas0fd00m/rfcat.git#egg=vstruct&subdirectory=vstruct
markup.py==1.12
